TVB-2631 Refactored AuthManager dependency
