+++ framework_tvb/tvb/interfaces/rest/server/decorators/rest_decorators.py
@@ -33,10 +33,11 @@ from functools import wraps
 from flask import current_app, request
 from flask.json import dumps
 from keycloak.exceptions import KeycloakError
+from tvb.core.services.authorization import AuthorizationManager
 from tvb.core.services.user_service import UserService
 from tvb.interfaces.rest.commons.strings import Strings
 from tvb.interfaces.rest.commons.exceptions import AuthorizationRequestException
-from tvb.interfaces.rest.server.security.authorization import AuthorizationManager, set_current_user
+from tvb.interfaces.rest.server.request_helper import set_current_user
 
 
 def _convert(obj):
